requirements.txt
requests
google-ads
psycopg2-binary
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
pandas-gbq>=0.17.0
python-dotenv
google-generativeai
pandas
redis
rq
